requirements.txt
# Python 3.13 Compatible Requirements
# Core Framework
fastapi>=0.120.2
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
pydantic>=2.10.0  # Python 3.13 compatible (has pre-built wheels)
pydantic-settings>=2.7.0  # Updated for pydantic 2.10+
aiofiles>=23.2.1

# Database
neo4j>=5.15.0

# Machine Learning & NLP
torch>=2.5.0  # Python 3.13 compatible
numpy>=1.26.0  # Python 3.13 compatible (1.24.3 fails on Py3.13)
pandas>=2.2.0  # Updated for better compatibility
sentence-transformers>=3.0.0  # Updated version
transformers>=4.40.0  # Updated for compatibility

# Document Processing
pypdf>=6.1.3
python-docx>=1.1.0

# HTTP & API
httpx>=0.25.2

# LLM Integration
ollama>=0.1.7
google-generativeai>=0.8.0
tiktoken>=0.5.2

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Development Tools
ruff>=0.12.12